@@ -10,25 +10,24 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^15.2.6",
+    "@angular/animations": "^16.0.1",
     "@angular/cdk": "^15.0.0",
-    "@angular/common": "^15.2.6",
-    "@angular/compiler": "^15.2.6",
-    "@angular/core": "^15.2.6",
-    "@angular/flex-layout": "^14.0.0-beta.41",
-    "@angular/forms": "^15.2.6",
+    "@angular/common": "^16.0.1",
+    "@angular/compiler": "^16.0.1",
+    "@angular/core": "^16.0.1",
+    "@angular/forms": "^16.0.1",
     "@angular/material": "^15.0.0",
-    "@angular/platform-browser": "^15.2.6",
-    "@angular/platform-browser-dynamic": "^15.2.6",
-    "@angular/router": "^15.2.6",
+    "@angular/platform-browser": "^16.0.1",
+    "@angular/platform-browser-dynamic": "^16.0.1",
+    "@angular/router": "^16.0.1",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.12.0"
+    "zone.js": "~0.13.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.2.5",
-    "@angular/cli": "~15.2.5",
-    "@angular/compiler-cli": "^15.2.6",
+    "@angular-devkit/build-angular": "^16.0.1",
+    "@angular/cli": "~16.0.1",
+    "@angular/compiler-cli": "^16.0.1",
     "@types/jasmine": "~4.3.0",
     "jasmine-core": "~4.5.0",
     "karma": "~6.4.0",
@@ -36,6 +35,6 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.0.0",
-    "typescript": "~4.8.2"
+    "typescript": "~5.0.4"
   }
 }
